move queryContract to contracts
