+++ src/services/contract/custom/useAdminResourcesQuery.ts
@@ -1,12 +1,12 @@
 import useSWR, { SWRResponse } from "swr";
 import { AdminResources } from "services/types";
 import { CW3Config } from "types/contracts";
-import { queryContract } from "services/juno/queryContract";
 import { useGetWallet } from "contexts/WalletContext";
 import { idParamToNum } from "helpers";
 import { isJunoAddress } from "schemas/tests";
 import { contracts } from "constants/contracts";
 import { adminRoutes, appRoutes } from "constants/routes";
+import { queryContract } from "../queryContract";
 import { apCWs } from "./constants";
 
 export function useAdminResourcesQuery(
