+++ src/services/contract/queryContract/genQueryPath.ts
